extract duplicate code into common module: 1. litentry/litmus/moonbase/rococo lib.rs: * variables: RuntimeBlockLength BlockHashCount
2. litentry/litmus/moonbase/rococo transaction_payment.rs:
* variables: TargetBlockFullness AdjustmentVariable MinimumMultiplier SlowAdjustingFeeUpdate
